requirements.txt
# Core RAG dependencies
chromadb>=0.5.0
tiktoken>=0.7.0

# Document processing
pypdf>=4.0.0
pymupdf>=1.24.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# OCR support
pytesseract>=0.3.10
pdf2image>=1.17.0

# LLM & Embeddings
ollama>=0.3.0

# Web server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
streamlit>=1.35.0

# Utils
python-dotenv>=1.0.0
